// src/lib/api/generated/model/friend.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Etechs Social Network
 * API contract cho social-network-ui (main API + message service).
 * OpenAPI spec version: 1.0
 */
import type { Author } from './author';

export interface Friend {
  id: string;
  user: Author;
  created_at: string;
}